relocate `RunwayModuleType` class
