+++ runway/core/components/__init__.py
@@ -3,5 +3,12 @@ from ._deploy_environment import DeployEnvironment
 from ._deployment import Deployment
 from ._module import Module
 from ._module_path import ModulePath
+from ._module_type import RunwayModuleType
 
-__all__ = ["DeployEnvironment", "Deployment", "Module", "ModulePath"]
+__all__ = [
+    "DeployEnvironment",
+    "Deployment",
+    "Module",
+    "ModulePath",
+    "RunwayModuleType",
+]
